Remove unecessary steps in image build process.
The image was wasting a lot of time importing, deleting and reimporting modules.
Changing to using the latest official python image will save time in having to install and set up
python.  There is a change in operating system (as the official python image uses debian),
but for testing, it shouldn't make a difference.

Finally, a simple readme was added to make it a) look more professional and b) let people
know how new images make thier way into dockerhub.